Change icons to be static components
