+++ src/components/icon/Success/index.js
@@ -1,5 +1,6 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
+import config from 'vtex-tachyons/config.json'
 
 class Success extends PureComponent {
   render() {
@@ -18,6 +19,12 @@ class Success extends PureComponent {
       </svg>
     )
   }
+}
+
+Success.defaultProps = {
+  fill: config.colors['serious-black'],
+  width: 16,
+  height: 16,
 }
 
 Success.propTypes = {
