+++ styleguide.config.js
@@ -18,7 +18,6 @@ module.exports = {
           './src/components/Alert/index.js',
           './src/components/Button/index.js',
           './src/components/Card/index.js',
-          './src/components/Icon/index.js',
           './src/components/Input/index.js',
           './src/components/Spinner/index.js',
           './src/components/Toggle/index.js',
